Set up pylint RC file
This is the right place to add custom rules about what is okay. In
particular I think we may need to add something there to suppress all
those warnings about astropy units; it would also be good to turn on
variable name warnings, but we'd need to let it know that variables like
dX and i are kosher.