Massive amounts of changes to integrate astropy tables.
The TOAs class is now fully based around them, and a table is the main
thing that is produced.  The tables seem to pickle well, but can't
yet be written to FITS or HDF5.  Lots of code cleanup as well.  Things
are slower than they were with Jing's first table tests, but that is
because the phase calculations are not using the tables yet.  And I
still need to check on some correctness of the observatory positions
and velocities -- things seem a tiny bit off from before, but I need
to verify that.  The speeding-up of the phase calculations should be
quite easy with all of this stuff in place.  Note that mpmath is
almost completely purged as well -- I think it is only still there in
the spin-phase calculation now.  Maybe we can get rid of that too.