Update observatories.py
The clock corrections were read with a numpy.loadtxt call, and this only works if the clock file has no unexpected entries. Alas, in the Parkes file and maybe in others there are empty clock values and other problems, and so it crashes. 
If we stick to TEMPO-style clock files, we should tell the clock file reader the exact number of digits to read into the float values. This is a quick and dirty solution; there are probably faster ways you can imagine, in which case feel free to discard this pull request and change it with your preferred solution.