Implemented chi-squared calculation for correlated TOA errors
This re-uses the GLS code, so the GLS code has some slightly strange
options. Also to save computation and prevent infinite recursion the
chi2 attribute of a residuals object is now a property that calculates
the value on first request (and again on first request after update)
and caches the result.