Declarative setup.cfg; use tox.
Almost everything has been moved out of setup.py into setup.cfg; the
former can only be inspected by running it but the latter can be read
with no surprising dependencies or confusion.

To make sure that the installation includes everything necessary, to
actually use PINT, run tox. This builds a (source) distribution package,
creates an empty virtualenv, and installs the distribution package into
the fresh virtualenv with pip. This pulls in all declared dependencies
(and nothing else), and has access to only the files that are included in
the distribution package. Tox then switches to the tests directory (so
that ``import pint`` picks up only the installed package not ``pint/``)
and runs the test suite.