Implement working mjd_string, pulsar_mjd_string
The only PINT precision bug remaining is that ERFA behaves badly when
handed ancient MJDs. This could be fixed by array hackery that treats
any old elements as not being pulsar_mjd; there are no leap seconds
back then anyway. But writing this in a reliable way for
arbitrary-dimensional arrays is not trivial, and the result would be of
purely completionist value.