Work around np.longdouble bug.
Unfortunately, it seems np.longdouble(string) looses precision
if string is a numpy string scalar or array rather than a python
string.  This adjusts the test to not avoid that bug, but does
not actually check whether the bug introduces problems elsewhere
in the code.

See https://github.com/numpy/numpy/issues/15608